% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{number_tweets}
\alias{number_tweets}
\title{User Tweet Number Chart}
\usage{
number_tweets(username)
}
\arguments{
\item{username}{Twitter handle}
}
\value{

}
\description{
User Tweet Number Chart
}
\examples{
number_tweets("realDonaldTrump")
}
