% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{search_hashtag}
\alias{search_hashtag}
\title{Twitter Hashtag Data}
\usage{
search_hashtag(hashtag, n)
}
\arguments{
\item{hashtag}{Hashtag to search}

\item{n}{Number of tweets to retrieve}
}
\value{

}
\description{
Twitter Hashtag Data
}
\examples{
search_hashtag("data science", 3200)
}
