% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tweet_emoji_chart}
\alias{tweet_emoji_chart}
\title{Emoji Chart Plot}
\usage{
tweet_emoji_chart(username)
}
\arguments{
\item{username}{Twitter handle}
}
\value{

}
\description{
Emoji Chart Plot
}
\examples{
tweet_emoji_chart("realDonaldTrump")
}
