% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tweet_emojis}
\alias{tweet_emojis}
\title{Emoji Usage}
\usage{
tweet_emojis(username)
}
\arguments{
\item{username}{Twitter handle}
}
\value{

}
\description{
Emoji Usage
}
\examples{
tweet_emojis("realDonaldTrump")
}
