% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tweet_length}
\alias{tweet_length}
\title{User Tweet Length Chart}
\usage{
tweet_length(username)
}
\arguments{
\item{username}{Twitter handle}
}
\value{

}
\description{
User Tweet Length Chart
}
\examples{
tweet_length("realDonaldTrump")
}
