% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tweets_to_wordcloud}
\alias{tweets_to_wordcloud}
\title{User Wordcloud Generator}
\usage{
tweets_to_wordcloud(username)
}
\arguments{
\item{username}{Twitter handle}
}
\value{

}
\description{
User Wordcloud Generator
}
\examples{
tweets_to_wordcloud("gates_duncan")
}
