% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{user_tweets}
\alias{user_tweets}
\title{Twitter User Data}
\usage{
user_tweets(user, n)
}
\arguments{
\item{user}{User's handle}

\item{n}{Number of tweets to retrieve}
}
\value{

}
\description{
Twitter User Data
}
\examples{
user_tweets("realDonaldTrump", 3200)
}
